+++ umongo/embedded_document.py
@@ -3,6 +3,14 @@ from .data_objects import BaseDataObject
 from .data_proxy import missing
 
 
+__all__ = (
+    'EmbeddedDocumentTemplate',
+    'EmbeddedDocument',
+    'EmbeddedDocumentOpts',
+    'EmbeddedDocumentImplementation'
+)
+
+
 class EmbeddedDocumentTemplate(Template):
     """
     Base class to define a umongo embedded document.
@@ -69,6 +77,9 @@ class EmbeddedDocumentImplementation(Implementation, BaseDataObject):
     def clear_modified(self):
         self._modified = False
         self._data.clear_modified()
+
+    def required_validate(self):
+        self._data.required_validate()
 
     def from_mongo(self, data):
         self._data.from_mongo(data)