+++ HISTORY.rst
@@ -2,6 +2,14 @@
 History
 =======
 
+dev
+---
+
+* Document&EmbeddedDocument store fields passed during initialization
+  as modified (see #50)
+* Required field inside embedded document are handled correctly (see #61)
+* Document support marshmallow's pre/post processors
+
 0.10.0 (2016-09-29)
 -------------------